argumentos de tiempo se parsean en milisegundos

+++ philosophers.h
@@ -41,7 +41,7 @@ typedef struct	s_philo
 void	*ft_calloc(size_t nmemb, size_t size);
 void	ft_free_list(t_philo **start);
 void	ft_data_printer(t_philo *philo_d);
-int		ft_atoi(char *s);
+long	ft_atoi(char *s);
 int		ft_strlen(char *s);
 int		ft_parse(t_data *data, char **argv);
